requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.4.0
openai==1.76.0
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-docx==1.1.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
